+++ test/api.test.js
@@ -157,7 +157,12 @@ function makeApp(dataDir) {
     error: () => {},
     setPluginStatus: () => {},
     setPluginError: () => {},
-    savePluginOptions: (_options, callback) => callback && callback(),
+    savePluginOptions: (_options, callback) => {
+      if (typeof callback !== 'function') {
+        throw new TypeError('callback is not a function')
+      }
+      callback(null)
+    },
     getDataDirPath: () => dataDir,
     handleMessage: () => {},
     config: { port: 3000 },